+++ examples/documentation/client/awesome_client/main.py
@@ -2,5 +2,5 @@
 
 
 def main(value_1: int, value_2: int) -> int:
-    """ main function of awesome_client """
+    """main function of awesome_client"""
     return value_1 + value_2