+++ examples/documentation/client/tests/test_main.py
@@ -3,5 +3,5 @@ import awesome_client.main
 
 
 def test_main() -> None:
-    """ Tests for the main function """
+    """Tests for the main function"""
     assert awesome_client.main.main(3, 7) == 10